Add search option
I had feedback that it is not clear how to select a city if more than one with the same name exist in the same country or in any country.
I decided to add that functionality.
I fetched data from the API endpoint for all the cities with the same name and then I will add a function to extract the weather for this particular city the user chooses.
I have added the first part to the project and the cities are available to view.
I added some JSX for that and added a onclick event with the function.
I managed to get the API endpoint but will have to use geo coordinates as the json data does not contain ID for the city.