Error message
I fixed the error message as it was not displaying. I added a conditional JSX to show the message if the city array is empty.
I fixed a small issue with the state/province. So basically I show city, county and country but some cities do not show a county and there was an extra comma and empty space. I added a conditional that the county should be shown only if it has a value.
I also added a countrycode js file with a json file with all country codes with their respective countries. After that I render the whole country name on the screen.