Error message
I added some conditions for the error message as there was no error message shown if the user searches for an empty string or the string 'test'.
As the API does not return an error when incorrect data is searched for, I decided to have a condition to show 2 different error messages and gave them states. I used the state hook to show/hide the message the same way I did with the rest of the components.
Now there are 2 different messages based on the user input.